Modify model/EnumModel.py app.py
